Refactor folder names and file placements
