+++ src/screens/plateinventoryscreen/PlateInventoryScreen.tsx
